added implementation for ray calculation
